LayoutTests/fast/profiler/event-handler.html

<html>
<head>
<script src="resources/profiler-test-JS-resources.js"></script>
<script>
if (window.layoutTestController)
    layoutTestController.dumpAsText();

console.profile("Event handler");
function testStart()
{    
    var myButton = document.getElementById("buttonWithAnEventListener");
    myButton.addEventListener('click', insertNewText, false);
    myButton.click();

    endTest();
}
</script>
</head>

<body onload="testStart()">
This page has an event handler.
<br>
<br>
To use this test, load it in the browser then load the WebInspector and look at
the profile.  In the profile insertnewText() should be a sibling to onload().
<input type="button" id="buttonWithAnEventListener" value="Button with an event listener">
<div id="output"></div>
</body>
</html>
